Big INT

+++ db/migrate/20190321112427_create_products.rb
@@ -9,7 +9,7 @@ class CreateProducts < ActiveRecord::Migration::Current
       create_table table, options do |t|
         t.string      :vad_variant_code
         t.string      :vad_description
-        t.integer     :vad_ean_code
+        t.bigint      :vad_ean_code
         t.integer     :free_stock
         t.integer     :on_order
         t.date        :eta
